Rename parser_v* to cwl_v*
