+++ cwl_utils/cwl_v1_2_expression_refactor.py
@@ -27,7 +27,7 @@ from ruamel import yaml
 from schema_salad.sourceline import SourceLine
 from schema_salad.utils import json_dumps
 
-import cwl_utils.parser.parser_v1_2 as cwl
+import cwl_utils.parser.cwl_v1_2 as cwl
 
 
 def expand_stream_shortcuts(process: cwl.CommandLineTool) -> cwl.CommandLineTool:
