+++ cwl_utils/docker_extract.py
@@ -6,7 +6,7 @@ import sys
 from pathlib import Path
 from typing import Iterator, Union, cast
 
-import cwl_utils.parser.parser_v1_0 as cwl
+import cwl_utils.parser.cwl_v1_0 as cwl
 from cwl_utils.image_puller import (
     DockerImagePuller,
     ImagePuller,
