+++ tests/test_etools_to_clt.py
@@ -11,9 +11,9 @@ from _pytest.tmpdir import TempPathFactory
 from cwltool.errors import WorkflowException
 from pytest import raises
 
-import cwl_utils.parser.parser_v1_0 as parser
-import cwl_utils.parser.parser_v1_1 as parser1
-import cwl_utils.parser.parser_v1_2 as parser2
+import cwl_utils.parser.cwl_v1_0 as parser
+import cwl_utils.parser.cwl_v1_1 as parser1
+import cwl_utils.parser.cwl_v1_2 as parser2
 from cwl_utils.cwl_v1_0_expression_refactor import traverse as traverse0
 from cwl_utils.cwl_v1_1_expression_refactor import traverse as traverse1
 from cwl_utils.cwl_v1_2_expression_refactor import traverse as traverse2
