+++ tox.ini
@@ -47,7 +47,7 @@ commands =
   py{36,37,38,39}-unit: coverage xml
   py{36,37,38,39}-unit: codecov --file coverage.xml
   py{36,37,38,39}-lint: flake8 setup.py cwl_utils
-  py{36,37,38,39}-lint: black --diff --check --exclude 'parser_v.*' setup.py cwl_utils
+  py{36,37,38,39}-lint: black --diff --check --exclude 'cwl_v.*' setup.py cwl_utils
   py{36,37,38,39}-mypy: make mypy
 
 whitelist_externals =
